Fix type name uint->size_t. The old code doesn't compile on Windows.
BUG=None
TEST=win_aura passes the compile phase

Review URL: https://chromiumcodereview.appspot.com/11618027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@174243 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 29d642f3d696a6d9cc291a35d10d9c9242406099

+++ ui/message_center/notification_list_unittest.cc
@@ -119,7 +119,7 @@ class NotificationListTest : public testing::Test {
 
   scoped_ptr<MockNotificationListDelegate> delegate_;
   scoped_ptr<NotificationList> notification_list_;
-  uint counter_;
+  size_t counter_;
 
   DISALLOW_COPY_AND_ASSIGN(NotificationListTest);
 };
